COOL-15551 double instead of long

+++ projects/yuuvis/framework/src/lib/bpm/task-details/task-details-task/task-details-task.component.ts
@@ -125,7 +125,7 @@ export class TaskDetailsTaskComponent implements OnInit {
         break;
       }
       case 'decimal': {
-        pv.type = 'long';
+        pv.type = 'double';
         pv.value = formValue;
         break;
       }
